requirements.txt
torch
torchvision
numpy
imageio
pillow
matplotlib
gradio
